+++ lib/adaptive_platform_ui.dart
@@ -48,6 +48,7 @@ export 'src/widgets/adaptive_slider.dart';
 export 'src/widgets/adaptive_segmented_control.dart';
 export 'src/widgets/adaptive_alert_dialog.dart';
 export 'src/widgets/adaptive_popup_menu_button.dart';
+export 'src/widgets/adaptive_context_menu.dart';
 export 'src/widgets/adaptive_scaffold.dart';
 export 'src/widgets/adaptive_app_bar_action.dart';
 export 'src/widgets/adaptive_snackbar.dart';